backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0
beanie==1.23.6  # MongoDB ODM
certifi==2023.11.17  # SSL certificates for MongoDB Atlas

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
argon2-cffi==23.1.0

# Cloud Storage
boto3==1.34.0  # For Cloudflare R2 (S3-compatible)
botocore==1.34.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9
redis==5.0.1

# Email (for notifications)
fastapi-mail==1.4.1
resend==0.6.0
